% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{alameda_wide}
\alias{alameda_wide}
\title{Alameda county tract demographics}
\format{A data frame with columns:
\describe{
\item{fips}{U.S. tract FIPS code}
\item{county}{Name of county}
\item{total_pop_2010}{Gross value of maple products in thousands of Canadian dollars.}
\item{etc}{Columns by race/ethnicity. All races except hispanic exclude hispanics (i.e. white_2010 is the subset of non-hispanic white people.)}
}}
\source{
U.S. Census Bureau, 2010 tables P005001, P005003:P005009. Compiled with tidycensus prackage.
}
\usage{
alameda_wide
}
\description{
Data by race for each tract in Alameda County, Californai.
}
\examples{
\dontrun{
 alameda_wide
}
}
\keyword{datasets}
