% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bay_race}
\alias{bay_race}
\title{bay county tract demographics}
\format{
A data frame with columns:
\describe{
\item{fips}{U.S. tract FIPS code}
\item{total_pop}{Total tract population in 2010}
\item{county}{County name and year}
\item{asian,black,hispanic,white,all_other}{Columns by race/ethnicity. All races except hispanic exclude hispanics (i.e. white_2010 is the subset of non-hispanic white people.)}
}
}
\source{
U.S. Census Bureau, 2010 tables P005001, P005003:P005006, P005010. Compiled with tidycensus prackage.
}
\usage{
bay_race
}
\description{
Data by race for each tract in bay County, Californai.
}
\examples{
\dontrun{
 bay_race
}
}
\keyword{datasets}
