% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{detroit_race}
\alias{detroit_race}
\title{Detroit Metro Area tract demographics}
\format{
A data frame with columns:
\describe{
\item{tract}{U.S. tract FIPS code}
\item{population}{Total tract population in 2010}
\item{place_name}{Place name}
\item{asian,black,hispanic,white,other,nhpi, natam}{Columns by race/ethnicity. All races except hispanic exclude hispanics (i.e. white_2010 is the subset of non-hispanic white people.)}
}
}
\source{
U.S. Census Bureau, 2010 tables P005001, P005003:P005006, P005010. Compiled with tidycensus prackage.
}
\usage{
detroit_race
}
\description{
Data by race for each tract in the Detroit Metro Area. Tracts allocated to place using centroids.
}
\examples{
\dontrun{
 bay_race
}
}
\keyword{datasets}
