% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seg_measures.R
\name{dissimilarity}
\alias{dissimilarity}
\title{Dissimilarity Score}
\usage{
dissimilarity(group1, group2, .sum = TRUE)
}
\arguments{
\item{group1, group2}{Numeric vectors representing the popuplation of the
groups to be compared.}

\item{.sum}{If TRUE, will return one value. (Or one value per group if specifying
\code{dplyr::group_by}.) If FALSE, will return a vector equaling the length
of the input vectors.}
}
\value{
A vector the length of group1 & group2
}
\description{
Stuff
}
