% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_group_measures.R
\name{divergence_sanity}
\alias{divergence_sanity}
\title{Sanity checks and warnings for divergence}
\usage{
divergence_sanity(df, totalPop)
}
\description{
Sanity checks and warnings for divergence
}
