% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seg_measures.R
\name{exposure}
\alias{exposure}
\title{Exposure Index}
\usage{
exposure(group1, group2, totalPop)
}
\arguments{
\item{group1}{Numeric vectors representing the popuplation of the
groups to be compared.}

\item{group2}{Numeric vectors representing the popuplation of the
groups to be compared.}

\item{totalPop}{The total population of the geography (e.g. Census
tract) being analyzed. If not specified, deaults to the sum of the
populations provided in \code{...}}
}
\description{
Exposure Index
}
