% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seg_measures.R
\name{scale01}
\alias{scale01}
\title{Helper function to create scale from 0-100}
\usage{
scale01(x, minimum = max(x, na.rm = T), maximum = max(x, na.rm = T))
}
\description{
Helper function to create scale from 0-100
}
