% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seg_measures.R
\name{segregation_measures}
\alias{segregation_measures}
\alias{dissim}
\alias{exposure}
\alias{isolation}
\alias{location_quotient}
\alias{entropy}
\alias{entropy_index}
\alias{entropy_score}
\alias{divergence}
\alias{scale01}
\title{Segregation measures}
\usage{
dissim(race1, race2)

exposure(race1, race2, totalPop)

isolation(race, totalPop)

location_quotient(race, totalPop)

entropy(..., totalPop = NULL, scaled = FALSE, thresholds = FALSE)

entropy_index(entropy_smallGeo, entropy_bigGeo)

entropy_score(entropy_index, totalPop)

divergence(..., totalPop = NULL)

scale01(x, minimum = max(x, na.rm = T), maximum = max(x, na.rm = T))
}
\arguments{
\item{totalPop}{The total population in the smallest geography being used
(usually tracts).

\describe{
\item{\code{entropy}}{Entropy score (Ei). \verb{Ei = sum(Xim * log(1/Xim)} where Xim is the
proportion of racial group within the geography. }
\item{\code{entropy_score}}{Calculates the value of H (entropy index) for
large-scale geography. }
\item{\code{zipped_shapefile}}{Extract shapefile directory and shapefile name}
\item{\code{filepaths}}{Creates filenames for variables used within package for easy reading
across years. Returns variables for reading when testing package functions; if used
externally, they will be used in the global environment}
\item{\code{access_file}} {Wrapper function that simplifies saving files
into the proper directory.}
\item{\code{censusVar}} {Wrapper for \code{paste0} that simplifies calling the variable with
the correct year.}

}}
}
\description{
Six segregation measures
}
