% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decompose.R
\name{decompose_divergence}
\alias{decompose_divergence}
\title{Additively decompose divergence scores by and within groups}
\source{
Roberto, 2016. "A Decomposable Measure of Segregation and Inequality."
}
\usage{
decompose_divergence(
  dataframe,
  groupCol = class(dataframe),
  weights = "sum",
  output = "scores",
  ...
)
}
\arguments{
\item{dataframe}{A dataframe composed of numeric/integer columns, with the exception
of a column named in \code{groupCol}. All columns are used in the divergence calculation
except for those specified in \code{groupCol} and \code{weights}(optional)}

\item{groupCol}{Name of the column in the dataframe used for grouping or a vector of groups
equalength the number of rows in dataframe. Ignored if passing a grouped_df
to \code{dataframe}.}

\item{weights}{Any of:
\describe{
\itemize{
\item \code{sum}(default) Sets \code{weights} to the rowwise sum of \code{dataframe}
\item A character string of the column to use as the column of weights in
\code{dataframe}.
\item A numeric vector the length of the number of rows of \code{dataframe}
}
}}

\item{...}{Population vectors for every group included in the divergence
calculation.}
}
\value{
A dataframe with one row if \code{summed == TRUE} or else one row for each
list or group in \code{dataframe}.

The dataframe will have three columns: 'within_divergence', equivalent to
\code{divergence()} for each dataframe or group in \code{dataframe};
'between_divergence', the divergence score of each group's demographics compared
to the full population; and \code{weights}, the sum of the weights for each group.
The sum of \code{decompose_divergence(...,summed = T)} should
equal the result of \code{divergence(...,summed = T)}
}
\description{
The Divergence Index is additively decomposable. This function allows for
splitting a population into groups of observations and calculating the
divergence score within those groups and between those groups.
}
\details{
The sum of the scores reported in \code{decompose_divergence} when setting summed==TRUE
should always be equal to the
}
\note{
The \code{divergence} parameters for each group are set to their defaults
unless explicitly noted above.
}
