% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decompose.R
\name{decompose_divergence}
\alias{decompose_divergence}
\title{Additively decompose divergence scores by and within groups}
\source{
Roberto, 2016. "A Decomposable Measure of Segregation and Inequality."
}
\usage{
decompose_divergence(
  dataframe,
  groupCol = NULL,
  weightCol = NA,
  output = "scores"
)
}
\arguments{
\item{dataframe}{A dataframe composed of numeric/integer columns, with the exception
of a column named in \code{groupCol}. All columns are used in the divergence calculation
except for those specified in \code{groupCol} and \code{weights}(optional)}

\item{groupCol}{Name of the column(s) in the dataframe used for grouping.
if passing a \code{grouped_df} to \code{dataframe}, this parameter is ignored. If using multiple
groups, divergence will be aggregated by all unique combinations of all groups, and compared
to the total datafame.}

\item{weightCol}{Either NA (default), which sets \code{weightCol} to the rowwise
sum of \code{dataframe}; or a character string of the column name in \code{dataframe}
that should be used for weightCol.}

\item{output}{Any of:
\describe{
\item{"scores"}{Default. The individual within and between divergence scores for each
row or group, plus the total score.}
\item{"weighted"}{As summed, but includinga column of weights to use if summing them both
across the entire dataset.}
\item{"percentage"}{One row for each entry(or group) as in "scores," but scaled so each
observation reports a percentage of the total score that would be reproted with "summed".}
\item{"all"}{The output from \code{summed}, \code{weighted}, and \code{percentage.}}
\item{"scaled"}{Not yet implemented. Re-scales divergence scores the divergence index to
have a range  of 0 to 1 by dividing by its maximum value for a given population. See details.}
}}
}
\value{
A dataframe as specified by the \code{output} parameter.

The dataframe will have three columns: 'within_divergence', equivalent to
\code{divergence()} for each dataframe or group in \code{dataframe};
'between_divergence', the divergence score of each group's demographics compared
to the full population; and \code{weightCol}, the sum of the weights for each group.
The sum of \code{decompose_divergence(...,summed = T)} should
equal the result of \code{divergence(...,summed = T)}
}
\description{
The Divergence Index is additively decomposable. This function allows for
splitting a population into groups of observations and calculating the
divergence score within those groups and between those groups.
}
\details{
The sum of the scores reported in \code{decompose_divergence} when setting summed==TRUE
should always be equal to the

Deomposing the divergence index allows users to simultatneously examine the segregation within
and between groups of a large geography. Furthermore, users can assess the percentage of
segregation coming from each group.

The \code{output} paramater "scaled" transforms the divergence index
it from an absolute to a relative measure of inequality and segregation, and negates
several of its desirable properties, including aggregation equivalence and independence.
(See Roberto, 2016)
}
\note{
The \code{divergence} parameters for each group are set to their defaults
unless explicitly noted above.
}
