% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two_group_measures.R
\name{dissimilarity}
\alias{dissimilarity}
\title{Dissimilarity Score}
\source{
Duncan, Otis Dudley, and Beverly Duncan. A Methodological Analysis of Segregation Indexes.” American Sociological Review, vol. 20, no. 2, 1955, pp. 210–217. JSTOR, www.jstor.org/stable/2088328. Accessed 14 Mar. 2020.
}
\usage{
dissimilarity(group1, group2, population = NA, summed = TRUE, na.rm = TRUE)
}
\arguments{
\item{group1, group2}{Numeric vectors representing the proportions of each group in each observation}

\item{population}{if \code{group1} and \code{group2} are proportions, this parameter gives the population weights.
Set to 1 to avoid a warning measure if using population}

\item{summed}{If TRUE, will return a single summary statistic. (Or one value per group if specifying
\code{dplyr::group_by}.) If FALSE (default), will return a vector equaling the length
of the input vectors. If 'weighted' (only for divergence and information theory), returns
a vector as in FALSE, but with pre-weighted values,
such that \verb{sum(divergence(..., summed = 'weighted))} is equivalent to
\code{divergence(..., summed = T)}.}

\item{na.rm}{logical. Should missing values (including NaN) be removed?
Used only if \code{summed} is set to TRUE.}
}
\value{
A scalar value, see note.
}
\description{
This index measures the evenness with which two mutually exclusive
groups are distributed across the geographic units that make up a
larger geographic entity; for example, the distribution of blacks and
whites across the census tracts that make up a metropolis.
Its minimum value is zero and its maximum value is 100.
}
\details{
The index of dissimilarity is calculated as:
\eqn{1/2 * \Sigma (group1/sum(group1) - group2/sum(group2))}
}
\note{
Setting summed == FALSE will return by-observation measures, but this
measure is not meant to be decomposed. These results are for verification
purposes only.
}
