% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_group_measures.R, R/two_group_measures.R
\name{divergence}
\alias{divergence}
\title{Divergence Index}
\source{
Created by Elizabeth Roberto: \url{https://arxiv.org/abs/1508.01167}

Created by Elizabeth Roberto: \url{https://arxiv.org/abs/1508.01167}
}
\usage{
divergence(..., totalPop = NULL, na.rm = TRUE, .sum = FALSE)

divergence(..., totalPop = NULL, na.rm = TRUE, .sum = FALSE)
}
\arguments{
\item{...}{columns to be included in the calculation of the index.}

\item{totalPop}{The total population of the geography (e.g. Census
tract) being analyzed. If not specified, deaults to the sum of the
populations provided in \code{...}}

\item{na.rm}{logical. Should missing values (including NaN) be removed?}

\item{.sum}{If TRUE, will return one value. (Or one value per group if specifying
\code{dplyr::group_by}.) If FALSE, will return a vector equaling the length
of the input vectors.}
}
\description{
Calculates the divergence index of segregation

Calculates the divergence index of segregation
}
\examples{
divergence(alameda_wide$white,alameda_wide$hispanic,alameda_wide$asian,
alameda_wide$black, totalPop = alameda_wide$total_pop)

\dontrun{
# Entering dataframe will cause an error
divergence(alameda_wide[c("white","black","asian","hispanic")])
}

divergence(alameda_wide$white,alameda_wide$hispanic,alameda_wide$asian,
alameda_wide$black, totalPop = alameda_wide$total_pop)

\dontrun{
# Entering dataframe will cause an error
divergence(alameda_wide[c("white","black","asian","hispanic")])
}

}
