% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_group_measures.R
\name{divergence}
\alias{divergence}
\title{Divergence Index}
\source{
Created by Elizabeth Roberto: \url{https://arxiv.org/abs/1508.01167}
}
\usage{
divergence(..., totalPop = NULL, na.rm = TRUE, .sum = FALSE)
}
\arguments{
\item{...}{columns to be included in the calculation of the index.}

\item{totalPop}{The total population of the geography (e.g. Census
tract) being analyzed. If not specified, deaults to the sum of the
populations provided in \code{...}}

\item{na.rm}{logical. Should missing values (including NaN) be removed?
Used only if \code{.sum} is set to TRUE.}

\item{.sum}{If TRUE, will return a single summary statistic. (Or one value per group if specifying
\code{dplyr::group_by}.) If FALSE, will return a vector equaling the length
of the input vectors.}

\item{percentage}{Are your supplying whole numbers (FALSE, default) or
percentages of the populatoin (TRUE)? If the latter, totalPop is}
}
\description{
FIX? divergence_pct Calculates the divergence index of segregation
}
\examples{
divergence(bay_race$white,bay_race$hispanic,bay_race$asian,
bay_race$black, totalPop = bay_race$total_pop)

\dontrun{
# Entering dataframe will cause an error
divergence(bay_race[c("white","black","asian","hispanic")])
}

}
