% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_group_measures.R
\name{entropy}
\alias{entropy}
\alias{entropy_index}
\alias{entropy_score}
\title{Theil's Index of Entropy}
\source{
Scale and threshold methodology from Holloway et al (2011):
\url{https://www.tandfonline.com/doi/abs/10.1080/00330124.2011.585080}

Theil, Henri. 1972. Statistical Decomposition Analysis.
}
\usage{
entropy(
  ...,
  totalPop = NULL,
  entropy_type = "index",
  scaled = FALSE,
  thresholds = FALSE,
  .sum = TRUE
)

entropy_index(entropy_smallGeo, entropy_bigGeo)

entropy_score(entropy_index, totalPop)
}
\arguments{
\item{...}{columns to be included in the calculation of the index.}

\item{totalPop}{The total population of the geography (e.g. Census
tract) being analyzed. If not specified, deaults to the sum of the
populations provided in \code{...}}

\item{entropy_type}{One of: \describe{
\item{score}{t index in wiki aka entropy score}
\item{index}{theil's h, akal entropy index, aka Theil's information theory index}
}}

\item{scaled}{Scale entropy scores from 0-1. See sources}

\item{thresholds}{Returns scores in from 1-3 (low-high diversity), based on the
following criteria: \describe{
\item{1}{Scaled entropy values less than or equal to 0.3707 or one group constitutes
more than 80 percent of the population}
\item{2}{All observations not coded as 1 or 3}
\item{3}{Scaled entropy greater than 0.7414 AND no group constituting more than
45 percent of the population}
} See sources.}

\item{.sum}{If TRUE, will return one value. (Or one value per group if specifying
\code{dplyr::group_by}.) If FALSE, will return a vector equaling the length
of the input vectors.}

\item{entropy_smallGeo, entropy_bigGeo}{The small (e.g. tract, row) and large
(e.g. county, group) entries.}

\item{entropy_index}{depreciated. delete}
}
\value{
A single value if .sum==TRUE, or a vector equaling the length of the inputs. Note that if
\code{entropy_type} == "index", and .sum is FALSE, then the returned vector will be entropy index, unweighted by
population
}
\description{
Entropy is used to measure the the extent to which multiple distributions conform to
a baseline. In standard entropy (\code{entropy()}), the baseline is constant evenness. In
Theil's \emph{T} Index
}
\details{
\describe{
\item{\code{entropy}}{Entropy score (Ei). \deqn{Ei = \Sigma (X_{im} * ln(1/X_{im})}{Ei = \Sigma (Xim \* ln(1/Xim))}
where Xim is the
proportion of racial group within the geography. }
\item{\code{entropy_score}}{Calculates the value of H (entropy index) for
large-scale geography. }
}
}
\seealso{
\url{https://en.wikipedia.org/wiki/Theil_index}

#' @examples
entropy(alameda_wide$white,alameda_wide$hispanic,alameda_wide$asian,
alameda_wide$black, totalPop = alameda_wide$total_pop)
}
