% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seg_measures.R
\name{entropy}
\alias{entropy}
\alias{entropy_index}
\alias{entropy_score}
\title{Theil's Index of Entropy}
\usage{
entropy(..., totalPop = NULL, scaled = FALSE, thresholds = FALSE)

entropy_index(entropy_smallGeo, entropy_bigGeo)

entropy_score(entropy_index, totalPop)
}
\arguments{
\item{...}{columns to be included in the calculation of the index.}

\item{totalPop}{The total population of the geography (e.g. Census
tract) being analyzed. If not specified, deaults to the sum of the
populations provided in \code{...}}
}
\description{
\describe{
\item{\code{entropy}}{Entropy score (Ei). \verb{Ei = sum(Xim * log(1/Xim)} where Xim is the
proportion of racial group within the geography. }
\item{\code{entropy_score}}{Calculates the value of H (entropy index) for
large-scale geography. }
}
}
