% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_group_measures.R
\name{entropy}
\alias{entropy}
\title{Theil's Index of Entropy}
\source{
Theil, Henri. 1972. Statistical Decomposition Analysis.
}
\usage{
entropy(
  ...,
  population = NA,
  comparison = NULL,
  entropy_type = "entropy",
  logBase = exp(1),
  scale = FALSE,
  summed = FALSE,
  na.rm = TRUE,
  weights = NA,
  sumPercent = NA
)
}
\arguments{
\item{...}{Vectors containing proportions of each group, or a dataframe containing only those vectors.}

\item{population}{A vector of population totals for each row, or weights summing to a total of
one. If NA, will assume all populations/weights are equal. Set to 1 to silence warning.
If a string, will use the string as the named column of the dataframe provided in ...}

\item{comparison}{A vector of percentages that must equal the length of the
number of vectors or columns in \code{...}, representing the percentages of each
group in the larger (comparison) geography.}

\item{entropy_type}{One of: \describe{
\item{"entropy"}{Default.
\eqn{E_{i} = \Sigma (X_{im} * ln(1/X_{im})}{Ei = \Sigma (Xim \* ln(1/X_{im}))}
where Xim is the proportion of racial group within the geography i.}
\item{"information_theory"}{Theil's information theory index}
\item{"overall_entropy"}{Overall entropy for the summarized dataset. Reports one score for the
entire dataset, as when setting \code{summed}=TRUE.}
}}

\item{logBase}{Specify the base for the logathirm used in the equation. Natural logarithm by default.}

\item{scale}{Scale entropy scores from 0-1. Setting scale to TRUE
ignores the entropy_type (set to "entropy") and summed parameters.}

\item{summed}{If TRUE, will return a single summary statistic. (Or one value per group if specifying
\code{dplyr::group_by}.) If FALSE (default), will return a vector equaling the length
of the input vectors. If 'weighted' (only for divergence and information theory), returns
a vector as in FALSE, but with pre-weighted values,
such that \verb{sum(divergence(..., summed = 'weighted))} is equivalent to
\code{divergence(..., summed = T)}.}

\item{na.rm}{logical. Should missing values (including NaN) be removed?
Used only if \code{summed} is set to TRUE.}

\item{weights}{deprecated, use population.}

\item{sumPercent}{deprecated, will throw error.}
}
\value{
A single value if summed==TRUE, or a vector equaling the length of the inputs. Note that if
\code{entropy_type} == "index", and summed is FALSE, then the returned vector will be entropy index, unweighted by
population
}
\description{
Entropy is used to measure the the extent to which multiple distributions conform to
a baseline.
}
\examples{
library(rsegregation)
entropy(bay_race$white,bay_race$hispanic,bay_race$asian,
bay_race$black, bay_race$all_other, population = bay_race$total_pop)

}
\seealso{
\url{https://en.wikipedia.org/wiki/Generalized_entropy_index}
}
