% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_group_measures.R
\name{entropy}
\alias{entropy}
\title{Theil's Index of Entropy}
\source{
Theil, Henri. 1972. Statistical Decomposition Analysis.
}
\usage{
entropy(
  ...,
  totalPop = NULL,
  entropy_type = "index",
  scaled = FALSE,
  .sum = TRUE,
  na.rm = TRUE
)
}
\arguments{
\item{...}{columns to be included in the calculation of the index.}

\item{totalPop}{The total population of the geography (e.g. Census
tract) being analyzed. If not specified, deaults to the sum of the
populations provided in \code{...}}

\item{entropy_type}{One of: \describe{
\item{score}{t index in wiki aka entropy score}
\item{index}{theil's h, akal entropy index, aka Theil's information theory index}
}}

\item{scaled}{Scale entropy scores from 0-1. Setting scaled to TRUE
ignores the entropy_type and .sum parameters}

\item{.sum}{If TRUE, will return a single summary statistic. (Or one value per group if specifying
\code{dplyr::group_by}.) If FALSE, will return a vector equaling the length
of the input vectors.}

\item{na.rm}{logical. Should missing values (including NaN) be removed?
Used only if \code{.sum} is set to TRUE.}
}
\value{
A single value if .sum==TRUE, or a vector equaling the length of the inputs. Note that if
\code{entropy_type} == "index", and .sum is FALSE, then the returned vector will be entropy index, unweighted by
population
}
\description{
Entropy is used to measure the the extent to which multiple distributions conform to
a baseline. In standard entropy (\code{entropy()}), the baseline is constant evenness. In
Theil's \emph{T} Index
}
\details{
\describe{
\item{\code{entropy}}{Entropy score (Ei). \eqn{E_{i} = \Sigma (X_{im} * ln(1/X_{im})}{Ei = \Sigma (Xim \* ln(1/X_{im}))}
where Xim is the
proportion of racial group within the geography. }
\item{\code{entropy_score}}{Calculates the value of H (entropy index) for
large-scale geography. }
}
}
\seealso{
\url{https://en.wikipedia.org/wiki/Theil_index}

#' @examples
entropy(bay_race$white,bay_race$hispanic,bay_race$asian,
bay_race$black, totalPop = bay_race$total_pop)
}
