% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_group_measures.R
\name{entropy}
\alias{entropy}
\title{Theil's Index of Entropy}
\source{
Theil, Henri. 1972. Statistical Decomposition Analysis.
}
\usage{
entropy(
  ...,
  weights = "sum",
  sumPercent = "weights",
  entropy_type = "entropy",
  scale = FALSE,
  summed = FALSE,
  na.rm = TRUE
)
}
\arguments{
\item{...}{Population vectors for every group included in the divergence
calculation.}

\item{weights}{Only used if \code{summed} or \code{weighted} is set to \code{TRUE}.
This can be either a vector of weights summing to one, or the total
population for each observation in \code{...}. Can be any of:
\describe{
\item{A numeric vector}{The length of each vector provided in each
entry in \code{...} (i.e. the column-wise length of \code{...}). Can be either population
or weights by observation}
\item{\code{sum}}{Sets \code{weights} to the rowwise sum of \code{...}}
\item{\code{none}}{Weighs each observation evenly (1/length). Note that if \code{...} is a
set of percentages summing to 1 for each row, \code{sum} and \code{none} are equivalent.}
}}

\item{sumPercent}{The percentage of each group in the larger population (i.e. the
population composed of the sum of groups provided in \code{...}). Can be any of:
\describe{
\item{A numeric vector}{Each entry represents the total population of the  with one entry for each group/vector provided in \code{...}. (i.e.
the rowwise length of \code{...}).
Note that for this to work correctly, each group must be provided in the same order in
\code{sumPercent} as in \code{...} Useful if the sum of \code{sumPercent} is less than 100\%, as \code{divergence}
will otherwise force}
\item{\code{weights}}{Default. Uses the value of \code{weights} to construct total population proportions.
If weights is set to \code{none}, the total population proportions will be the unweighted
average of the percentages in each observaiton.}

}}

\item{entropy_type}{One of: \describe{
\item{"entropy"}{Default. t index in wiki aka entropy score}
\item{"information_theory"}{Theil's information theory index}
}}

\item{scale}{Scale entropy scores from 0-1. Setting scale to TRUE
ignores the entropy_type and summed parameters}

\item{summed}{If TRUE, will return a single summary statistic. (Or one value per group if specifying
\code{dplyr::group_by}.) If FALSE, will return a vector equaling the length
of the input vectors.}

\item{na.rm}{logical. Should missing values (including NaN) be removed?
Used only if \code{summed} is set to TRUE.}
}
\value{
A single value if summed==TRUE, or a vector equaling the length of the inputs. Note that if
\code{entropy_type} == "index", and summed is FALSE, then the returned vector will be entropy index, unweighted by
population
}
\description{
Entropy is used to measure the the extent to which multiple distributions conform to
a baseline.
}
\details{
\describe{
\item{\code{entropy}}{Entropy score (Ei). \eqn{E_{i} = \Sigma (X_{im} * ln(1/X_{im})}{Ei = \Sigma (Xim \* ln(1/X_{im}))}
where Xim is the
proportion of racial group within the geography. }
\item{\code{entropy_score}}{Calculates the value of H (entropy index) for
large-scale geography. }
}
}
\examples{
library(rsegregation)
entropy(bay_race$white,bay_race$hispanic,bay_race$asian,
bay_race$black, bay_race$all_other, weights = bay_race$total_pop)

}
\seealso{
\url{https://en.wikipedia.org/wiki/Theil_index}
}
