% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two_group_measures.R
\name{exposure}
\alias{exposure}
\title{Exposure Index}
\source{
Wendell Bell, A Probability Model for the Measurement of Ecological Segregation, Social Forces, Volume 32, Issue 4, May 1954, Pages 357–364, https://doi.org/10.2307/2574118
}
\usage{
exposure(group1, group2, totalPop, summed = TRUE, na.rm = TRUE)
}
\arguments{
\item{group1}{Numeric vectors representing the popuplation of the
groups to be compared.}

\item{group2}{Numeric vectors representing the popuplation of the
groups to be compared.}

\item{totalPop}{A numeric vector the length of \code{group} with overall population totals.}

\item{summed}{If TRUE, will return a single summary statistic. (Or one value per group if specifying
\code{dplyr::group_by}.) If FALSE (default), will return a vector equaling the length
of the input vectors. If 'weighted' (only for divergence and information theory), returns
a vector as in FALSE, but with pre-weighted values,
such that \verb{sum(divergence(..., summed = 'weighted))} is equivalent to
\code{divergence(..., summed = T)}.}

\item{na.rm}{logical. Should missing values (including NaN) be removed?
Used only if \code{summed} is set to TRUE.}
}
\value{
A scalar value, see note.
}
\description{
The summation is over all the geographic units, e. g. census tracts,
comprising the larger geographic entity for which the exposure index
is being calculated.
}
\details{
The maximum value of the exposure index is the percent in the second
group.  That is, if blacks make up 30 percent of the population of a
metropolis, the maximum value of the average percent black for white
residents of that metropolis will be 30 percent.  This will require
that the index of dissimilarity measuring the evenness with which
blacks and every other racial group are distributed across the
metropolis equals zero. The minimum value of the exposure index is zero.  That is, although blacks might make up 30 percent of a metropolis, whites could live in exclusively white neighborhoods.   If the exposure index equals zero, than the index of dissimilarity comparing those two groups will equal zero. The exposure index involves two mutually exclusive racial groups.  However, the average percent black in the census tract of the typical white in a metropolis is almost always different from the average percent white for the typical black living in the same metropolis.

The formula for exposure is \eqn{\Sigma (g1/G1)*(g2/t)},
where g1=group1, G1=sum(group1),g2=group2, and t=totalPop
}
\note{
Setting summed == FALSE will return by-observation measures, but this
measure is not meant to be decomposed. These results are for verification
purposes only.
}
