% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_group_measures.R
\name{isolation}
\alias{isolation}
\title{Isolation Index}
\source{
Wendell Bell, A Probability Model for the Measurement of Ecological Segregation, Social Forces, Volume 32, Issue 4, May 1954, Pages 357–364, https://doi.org/10.2307/2574118
}
\usage{
isolation(group, totalPop, .sum = TRUE, na.rm = TRUE)
}
\arguments{
\item{group}{A numeric vector of population}

\item{totalPop}{The total population of the geography (e.g. Census
tract) being analyzed. If not specified, deaults to the sum of the
populations provided in \code{...}}

\item{.sum}{If TRUE, will return a single summary statistic. (Or one value per group if specifying
\code{dplyr::group_by}.) If FALSE, will return a vector equaling the length
of the input vectors.}

\item{na.rm}{logical. Should missing values (including NaN) be removed?
Used only if \code{.sum} is set to TRUE.}
}
\value{
A scalar value, see note.
}
\description{
The summation is over all the component geographic parts of the larger
geographic entity for which the isolation index is calculated.
}
\details{
This will report the percentage of population white in the geographic
unit, e.g., tract, for the typical or average white person.
The maximum value of this isolation index is 100.
Even if whites make up only 20 percent of a metropolis population,
all of them could live in all-white neighborhoods.
The minimum value of the isolation index is asymptotically close to 0.
That is, if there is only one white person in a metropolis of 100,000,
he or she would live in a geographic unit in which the percent white
was close to zero.

The isolation formula is
\eqn{\sigma ((group/sum(group))*(group/totalPop))}
}
\note{
Setting .sum == FALSE will return by-observation measures, but this
measure is not meant to be decomposed. These results are for verification
purposes only.
}
