% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seg_measures.R
\name{isolation}
\alias{isolation}
\title{Isolation Index}
\usage{
isolation(group, totalPop)
}
\arguments{
\item{group}{A numeric vector of population}

\item{totalPop}{The total population of the geography (e.g. Census
tract) being analyzed. If not specified, deaults to the sum of the
populations provided in \code{...}}
}
\description{
A single-group measure of the degree to which a group is isolated in
the full dataset
}
