% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_group_measures.R
\name{location_quotient}
\alias{location_quotient}
\alias{lq}
\title{Location Quotient}
\source{
U.S. Bureau of Economic Analysis: https://www.bea.gov/help/faq/478
}
\usage{
location_quotient(group, totalPop, na.rm = FALSE)

lq(group, totalPop, na.rm = FALSE)
}
\arguments{
\item{group}{A numeric vector of subpopulation totals}

\item{totalPop}{A numeric vector the length of \code{group} with population totals.}

\item{na.rm}{logical. Should missing values (including NaN) be removed?
Used only if \code{summed} is set to TRUE.}
}
\description{
A location quotient (LQ) is an analytical statistic that measures a
region’s concentration of a particular attribut relative to a larger geographic
unit (e.g. census tract to state). In economics, an
LQ is computed as an industry’s share of
a regional total for some economic statistic (earnings, GDP by
metropolitan area, employment, etc.) divided by the industry’s share
of the national total for the same statistic. For example, an LQ of
1.0 in mining means that the region and the nation are equally
specialized in mining; while an LQ of 1.8 means that the region has a
higher concentration in mining than the nation.
}
\details{
LQ is equivalent to the percentage of a group in a single observation divided
by the percentage of that group in the population of all observations, i.e.
\eqn{ (group/totalPop)/(\Sigma(group)/\Sigma(totalPop))}
}
\examples{
data("bay_race")
# calculate the location quotient for Hispanic people in each tract in the bay area
location_quotient(group = bay_race$hispanic,
totalPop = bay_race$total_pop)

}
